requirements.txt
aiofiles==23.2.1
aiogram==3.2.0
aiohttp==3.9.1
aiosignal==1.3.1
annotated-types==0.6.0
async-timeout==4.0.3
attrs==23.1.0
binance-connector==3.5.1
certifi==2023.11.17
charset-normalizer==3.3.2
contourpy==1.2.0
cycler==0.12.1
fonttools==4.46.0
frozenlist==1.4.0
greenlet==3.0.1
idna==3.6
kiwisolver==1.4.5
magic-filter==1.0.12
matplotlib==3.8.2
mplfinance==0.12.10b0
multidict==6.0.4
numpy==1.26.2
packaging==23.2
pandas==2.1.3
Pillow==10.1.0
psycopg==3.1.14
pycryptodome==3.19.0
pydantic==2.5.2
pydantic_core==2.14.5
pyparsing==3.1.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
requests==2.31.0
six==1.16.0
SQLAlchemy==2.0.23
typing_extensions==4.8.0
tzdata==2023.3
urllib3==2.1.0
websocket-client==1.7.0
yarl==1.9.3
